wappalyzer/requirements.txt
python-Wappalyzer
